+++ examples/js/basic-example.js
@@ -1,3 +1,4 @@
+/*jslint browser: true */
 /*global tabOverride */
 
 // enable Tab Override for all textarea elements on the page