+++ examples/js/requirejs-example.js
@@ -1,7 +1,8 @@
+/*jslint browser: true */
 /*global require */
 
 // load Tab Override
-require(['../build/taboverride.min.js'], function (tabOverride) {
+require(['../build/output/taboverride.min.js'], function (tabOverride) {
     'use strict';
 
     var textarea = document.getElementById('txt');